feat: Change print function to display reduced THIR. Due to change print function make thir directory and moved printer.rs to there.
