+++ Makefile
@@ -1,3 +1,3 @@
 RUST_LIB_PATH=$(rustc --print target-libdir)
 
-cargo run "sample/let_x.rs" -L "$RUST_LIB_PATH"
+cargo run "sample/test.rs" -L "$RUST_LIB_PATH"
