+++ sample/test.rs
@@ -1,0 +1,4 @@
+fn test(n: i32) {
+  let m = n * n;
+  let x = assert!(m >= 0);
+}
