+++ src/util.rs
@@ -1,14 +1,4 @@
 // rustc crates
-// use rustc_middle::ty::TyCtxt;
 
 // std crates
 // Own crates
-
-// pub fn print_thir<'tcx>(tcx: TyCtxt<'tcx>) {
-//   tcx.mir_keys(()).iter().for_each(|&key| {
-//     match tcx.thir_body(key) {}
-//     let mut printer = ThirPrinter::new(&thir);
-//     printer.print();
-//     printer.into_buffer();
-//   });
-// }
